requirements.txt
Cheetah==2.4.4
Django==1.8.3
Landscape-Client==14.12
Markdown==2.6.2
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.10
South==1.0.1
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
Unidecode==0.04.18
Whoosh==2.7.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
django-admin-bootstrapped==2.3.1
django-chartit==0.1
django-classy-tags==0.6.2
django-email-extras==0.3.1
django-endless-pagination==2.0
django-exportdata==0.2
django-filter==0.10.0
django-fixture-magic==0.0.7
django-forms-builder==0.12.2
django-grappelli==2.6.3
django-haystack==2.4.0
django-haystackbrowser==0.6.0
django-mandrill==0.1.0
django-select2-forms==1.1.18
django-simple-open-graph==0.3
django-tables2==0.15.0
django-wysiwyg-redactor==0.4.8.1
djangorestframework==3.1.3
docopt==0.4.0
future==0.9.0
html5lib==0.999
jsonpatch==1.3
jsonpointer==1.0
mandrill==1.0.57
oauth==1.0.1
prettytable==0.7.2
pyOpenSSL==0.13
pycurl==7.19.3
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
python-gnupg==0.3.7
pytz==2015.4
requests==2.2.1
six==1.5.2
sphinx-me==0.3
ssh-import-id==3.21
urllib3==1.7.1
wsgiref==0.1.2
zope.interface==4.0.5
